requirements.txt
numpy
pandas
pyyaml
pytest
opencv-python
torch
